fix int test
